Changed devman/setpollinterval to devman/devconfig
Fixed some integer/unsigned warnings
Changed setDevicePollInterval to setDevicePollIntervalUs
getBusByName allows numbers
Removed getBusByNameOrNumberString

+++ components/core/FileSystem/FileSystem.cpp
@@ -722,7 +722,7 @@ SpiramAwareUint8Vector FileSystem::getFileSection(const String& fileSystemStr, c
     // Return data
     if (readLen <= 0)
         return SpiramAwareUint8Vector();
-    else if (readLen < fileData.size())
+    else if (readLen < (int)fileData.size())
         fileData.resize(readLen);
     return fileData;
 }
